requirements.txt
asgiref==3.7.2
Django==4.2.4
django-cors-headers==4.2.0
django-extensions==3.2.3
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
geographiclib==2.0
geopy==2.3.0
Pillow==10.0.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4
